strollin/requirements.txt
opencv-python
face_recognition
spotipy
python-dotenv
numpy
pytest
